Rename the Ethereum compilance test package
