+++ src/test/java/org/ethereum/vm/compliance/spec/Environment.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.ethereum.vm.test.api;
+package org.ethereum.vm.compliance.spec;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
